itinerary functions completed with authorisation, endpoint.json updated

+++ src/main/java/dev/group2/traveldiary/travel_diary_backend/controller/ItineraryController.java
@@ -1,9 +1,13 @@
 package dev.group2.traveldiary.travel_diary_backend.controller;
 import dev.group2.traveldiary.travel_diary_backend.model.Itinerary;
 import dev.group2.traveldiary.travel_diary_backend.dto.ItineraryDTO;
+import dev.group2.traveldiary.travel_diary_backend.model.User;
+import dev.group2.traveldiary.travel_diary_backend.repository.UserRepository;
 import dev.group2.traveldiary.travel_diary_backend.service.ItineraryService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Map;
@@ -12,13 +16,18 @@ import java.util.Map;
 @RequestMapping("/api/itineraries")
 public class ItineraryController {
     private final ItineraryService itineraryService;
+    private final UserRepository userRepository;
 
-    public ItineraryController(ItineraryService itineraryService) {
+    public ItineraryController(ItineraryService itineraryService, UserRepository userRepository) {
         this.itineraryService = itineraryService;
+        this.userRepository = userRepository;
     }
 
     @PostMapping
-    public ResponseEntity<ItineraryDTO> addItinerary(@RequestBody Itinerary itinerary) {
+    public ResponseEntity<Object> addItinerary(@AuthenticationPrincipal UserDetails userDetails,@RequestBody Itinerary itinerary) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(Map.of("message", "Unauthorized operation. Please login again."));
+        }
        Itinerary savedItinerary = itineraryService.addItinerary(itinerary);
        ItineraryDTO savedItineraryDTO = new ItineraryDTO(savedItinerary);
        return ResponseEntity.status(HttpStatus.CREATED).body(savedItineraryDTO);
@@ -43,14 +52,30 @@ public class ItineraryController {
     }
 
     @PatchMapping("/{itineraryId}")
-    public ResponseEntity<ItineraryDTO> modifyItinerary(@PathVariable Long itineraryId,
-                                                        @RequestParam(required = false) String title,
-                                                        @RequestParam(required = false) Boolean isPrivate){
-        return ResponseEntity.status(HttpStatus.OK).body(itineraryService.updateItinerary(itineraryId,title,isPrivate));
+    public ResponseEntity<Object> modifyItinerary(@PathVariable Long itineraryId,
+                                                        @RequestBody Itinerary itinerary,
+                                                        @AuthenticationPrincipal UserDetails userDetails){
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(Map.of("message", "Unauthorized operation."));
+        }
+        User user = userRepository.findByUsername(userDetails.getUsername()).orElseThrow();
+        Long itineraryUserId = itineraryService.getItineraryById(itineraryId).getUserId();
+        if(!itineraryUserId.equals(user.getUserId())){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(Map.of("message", "Unauthorized operation."));
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(itineraryService.updateItinerary(itineraryId,itinerary.getTitle(),itinerary.getIsPrivate()));
     }
 
     @DeleteMapping( "/{itineraryId}" )
-    public ResponseEntity<Map<String,String>> deleteItinerary(@PathVariable Long itineraryId){
+    public ResponseEntity<Map<String,String>> deleteItinerary(@PathVariable Long itineraryId,@AuthenticationPrincipal UserDetails userDetails){
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(Map.of("message", "Unauthorized operation."));
+        }
+        User user = userRepository.findByUsername(userDetails.getUsername()).orElseThrow();
+        Long itineraryUserId = itineraryService.getItineraryById(itineraryId).getUserId();
+        if(!itineraryUserId.equals(user.getUserId())){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(Map.of("message", "Unauthorized operation."));
+        }
         itineraryService.deleteItinerary(itineraryId);
         return ResponseEntity.status(HttpStatus.OK).body(Map.of("message","Itinerary Deleted Successfully"));
     }
